어떤 값을 어디에 줄 것인가
주문이란 개념 => OrderVO => 결제정보, 수령인(배송지)정보, 주문 상품 정보를 가진 것

+++ theTipsyPerfume/src/main/java/com/kh/ttp/productSale/billing/payment/model/vo/PaymentVO.java
@@ -1,7 +1,5 @@
 package com.kh.ttp.productSale.billing.payment.model.vo;
 
-import java.util.List;
-
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -16,7 +14,7 @@ import lombok.ToString;
 public class PaymentVO {
 
 	private int paymentNo;
-	private double paidAmount;
+	private int paidAmount;
 	
 	private String applyNum;
